+++ app/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/login-form"
+import { LoginForm } from "@/components/login-page/login-form"
 
 export default function LoginPage() {
   return (
