+++ components/home/home-body.tsx
@@ -1,0 +1,21 @@
+import {SectionCards} from "@/components/section-cards";
+import {ChartAreaInteractive} from "@/components/chart-area-interactive";
+import {DataTable} from "@/components/data-table";
+import data from "@/app/home/data.json";
+
+
+export function HomeBody() {
+    return (
+        <div className="flex flex-1 flex-col">
+            <div className="@container/main flex flex-1 flex-col gap-2">
+                <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
+                    <SectionCards/>
+                    <div className="px-4 lg:px-6">
+                        <ChartAreaInteractive/>
+                    </div>
+                    <DataTable data={data}/>
+                </div>
+            </div>
+        </div>
+    )
+}
