+++ components/login-page/login-form/dont-have-account-section.tsx
@@ -1,0 +1,11 @@
+
+export function DontHaveAccountSection() {
+    return (
+        <div className="text-center text-sm">
+            Don&apos;t have an account?{" "}
+            <a href="#" className="underline underline-offset-4">
+                Sign up
+            </a>
+        </div>
+    )
+}
