+++ components/login-page/login-form/email-section.tsx
@@ -1,0 +1,16 @@
+import {Label} from "@/components/ui/label";
+import {Input} from "@/components/ui/input";
+
+export function EmailSection() {
+    return (
+        <div className="grid gap-3">
+            <Label htmlFor="email">Email</Label>
+            <Input
+                id="email"
+                type="email"
+                placeholder="m@example.com"
+                required
+            />
+        </div>
+    )
+}
