+++ components/login-page/login-form/login-button.tsx
@@ -1,0 +1,10 @@
+import {Button} from "@/components/ui/button";
+
+
+export function LoginButton() {
+    return (
+        <Button type="submit" className="w-full">
+            Login
+        </Button>
+    )
+}
