+++ components/login-page/login-form/password-section.tsx
@@ -1,0 +1,19 @@
+import {Label} from "@/components/ui/label";
+import {Input} from "@/components/ui/input";
+
+export function PasswordSection() {
+    return (
+        <div className="grid gap-3">
+            <div className="flex items-center">
+                <Label htmlFor="password">Password</Label>
+                <a
+                    href="#"
+                    className="ml-auto text-sm underline-offset-2 hover:underline"
+                >
+                    Forgot your password?
+                </a>
+            </div>
+            <Input id="password" type="password" required/>
+        </div>
+    )
+}
