+++ components/login-page/login-form/terms-and-privacy-prompt.tsx
@@ -1,0 +1,11 @@
+
+
+export function TermsAndPrivacyPrompt() {
+    return (
+        <div
+            className="text-muted-foreground *:[a]:hover:text-primary text-center text-xs text-balance *:[a]:underline *:[a]:underline-offset-4">
+            By clicking continue, you agree to our <a href="#">Terms of Service</a>{" "}
+            and <a href="#">Privacy Policy</a>.
+        </div>
+    )
+}
