+++ components/login-page/login-form/welcome-message.tsx
@@ -1,0 +1,11 @@
+
+export function WelcomeMessage() {
+    return (
+        <div className="flex flex-col items-center text-center">
+            <h1 className="text-2xl font-bold">Welcome back</h1>
+            <p className="text-muted-foreground text-balance">
+                Login to your Acme Inc account
+            </p>
+        </div>
+    )
+}
